% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse: enables users to query the page for genes, miRNAs, SNPs, diseases, pathways, and ontology associated with PCOS.}
\usage{
browse(page, filter = NULL, search_qry = "")
}
\arguments{
\item{page}{A string representing which page will it retrieve the page from (i.e genes, miRNA, SNPs...).}

\item{filter}{A string to filter tables using certain subsets.}

\item{search_qry}{A string used to filter the tables according to a keyword.}
}
\value{
It returns a data frame with results filtered according to parameters.
}
\description{
Browse: enables users to query the page for genes, miRNAs, SNPs, diseases, pathways, and ontology associated with PCOS.
}
