% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_algorithm.R
\name{generateComorbidityAnalysis}
\alias{generateComorbidityAnalysis}
\title{Comorbidity Analysis Tool}
\usage{
generateComorbidityAnalysis(disease_list, algorithm)
}
\arguments{
\item{disease_list}{character vector containing diseases to be analyzed.}

\item{algorithm}{character vector containing algorithm to be perform the analysis.}
}
\value{
prints a heatmap containing comorbidity between input diseases.
}
\description{
This tool can be used to predict comorbidity based on shared genes and their ontologies,
uniqueness of the shared genes, and the network-based separation between them. For more details of
 the methodology refer to \href{https://www.nature.com/articles/s41598-020-71418-8}{PCOSKBR2: a database of genes, diseases, pathways, and networks associated with polycystic ovary syndrome}.
}
