% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathwayAnalysis.R
\name{generatePathwayTable}
\alias{generatePathwayTable}
\title{Pathway enrichment Analysis:  For selected diseases, this tool gives a list of enriched pathways.}
\usage{
generatePathwayTable(disease_list, dataset, database = "KEGG")
}
\arguments{
\item{disease_list}{character vector containing diseases to be analyzed.}

\item{dataset}{character containing the dataset. Can be either miRNA or Genes.}

\item{database}{character vector containing the database from which pathways will be extracted. It can either be "KEGG" or "Reactome".}
}
\value{

}
\description{
Pathway enrichment Analysis:  For selected diseases, this tool gives a list of enriched pathways.
}
