% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkAnalysis.R
\name{getNetworkAnalysis}
\alias{getNetworkAnalysis}
\title{Network Analysis}
\usage{
\code{getNetworkAnalysis(disease_list, dataset = c("miRNAs", "Genes")}
}
\arguments{
\item{disease_list}{Diseases for which comorbidity with PCOS will be analyzed. A possible list of diseases can be retrieved using the \code{listDiseases} function.}

\item{dataset}{Dataset. It can either be "miRNAs" or "Genes".}
}
\value{
Disease-Disease Network.
}
\description{
Iluustrates a disease-disease network through shared genes. Each node represents a disease. A connection is made between two nodes is done if they share one or more genes.
Size of the node represents the number of genes associated with the given disease. The width of each edge increases with a higher number of shared genes.
}
\examples{
\code{getNetworkAnalysis(disease_list = c("Keratomalacia","Xerotic keratitis"), dataset = "Genes")}
}
\author{
Omar Hassoun
}
