% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getPCOS}
\alias{getPCOS}
\title{Imports the table from the PCOSKB website}
\usage{
getPCOS()
}
\value{
the table in the PCOSKB website but in the form of a dataframe
}
\description{
Imports the table from the PCOSKB website -in the form of a dataframe- into the R workspace.
}
