% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennAnalysis.R
\name{getVennAnalysis}
\alias{getVennAnalysis}
\title{Venn Analysis: This tool can be used to illustrate the unique and/or common genes, pathways and ontologies for 2 or more (up to 6) diseases.}
\usage{
getVennAnalysis(disease_list, option = "Genes")
}
\arguments{
\item{disease_list}{character vector containing diseases to be analyzed.}

\item{option}{a character containing string to be analyzed. It can either be a "Genes", "Pathways" or "Ontologies"}
}
\value{

}
\description{
Venn Analysis: This tool can be used to illustrate the unique and/or common genes, pathways and ontologies for 2 or more (up to 6) diseases.
}
