% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCAlgoriththms.R
\name{listCAlgorithms}
\alias{listCAlgorithms}
\title{Algorithm for Comorbidity Analysis}
\usage{
\code{listAlgorithms()}
}
\value{

}
\description{
Lists algorithm to be inputed in \code{generateComorbidityAnalysis()} function.
}
\examples{
\code{listAlgorithms()}
}
\author{
Omar Hassoun
}
