% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDiseaseGroup.R
\name{listDiseaseGroup}
\alias{listDiseaseGroup}
\title{List Disease Group}
\usage{
listDiseaseGroup(dataset = "Genes")
}
\arguments{
\item{dataset}{String containing either the miRNAs or Genes dataset. Genes is the default.}
}
\value{
character vector of disease groups in input dataset.
}
\description{
List Disease Group
}
