% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFilters.R
\name{listFilters}
\alias{listFilters}
\title{lisFilters}
\usage{
\code{listFilters(page)}
}
\arguments{
\item{page}{Page. A possible list of algorithms can be retrieved using the \code{listPages} function.}
}
\value{
A \code{vector} containing functions.
}
\description{
Lists or search filters available for a given page in PCOSKBR.
}
\examples{
\code{listFilters(page = "Genes")}
}
\author{
Omar Hassoun
}
