% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPages.R
\name{listPages}
\alias{listPages}
\title{listPages: Prints the pages in PCOSKBR2}
\usage{
listPages()
}
\value{
returns vector with pages you can search for in PCCOSKB.
}
\description{
listPages: Prints the pages in PCOSKBR2
}
