% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneNetworkAnalysis.R
\name{listTissues}
\alias{listTissues}
\title{listTissues}
\usage{
\code{listTissues()}
}
\value{
\code{vector} of all tissues.
}
\description{
Tissues used as input for \link[pcoskbR]{viewInteractions} function.
}
\examples{
\code{listTissues()}
}
