% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathwayAnalysis.R
\name{PathwayAnalysis}
\alias{PathwayAnalysis}
\title{Pathway enrichment Analysis}
\usage{
\code{PathwayAnalysis(disease_list, dataset = c("miRNAs", "Genes"), database = c("KEGG", "Reactome"))}
}
\arguments{
\item{disease_list}{Diseases for which comorbidity with PCOS will be analyzed. A possible list of diseases can be retrieved using the \code{listDiseases} function.}

\item{dataset}{Dataset. It can either be "miRNAs" or "Genes".}

\item{database}{Database from which pathways will be extracted. It can either be "KEGG" or "Reactome".}
}
\value{
\code{data.frame} of enriched pathways.
}
\description{
For selected diseases, this tool gives a list of enriched pathways.
}
\examples{
\code{PathwayAnalysis(disease_list = c("Tangier Disease", "Hypoalphalipoproteinemia", "Neuropathy"),
         dataset = "Genes",
         database = "KEGG")}
}
\author{
Omar Hassoun
}
