% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneNetworkAnalysis.R
\name{viewInteractions}
\alias{viewInteractions}
\title{View Interactions: Shows a network of interactions for an input gene.}
\usage{
viewInteractions(gene, tissue = "all")
}
\arguments{
\item{gene}{A character containing the gene name (symbol)}

\item{tissue}{character containing tissue type.}
}
\value{

}
\description{
View Interactions: Shows a network of interactions for an input gene.
}
