% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse: Retrieves information from the PCOSKB database}
\usage{
\code{browse(page = c("Genes", "miRNA", "Diseases", "miRNA Diseases", "SNPs", "Pathways", "Ontologies"),
 filter = NULL,
 search_qry = "")}
}
\arguments{
\item{page}{Page from which data is retrieved. A possible list of pages can be retreived using the \link[pcoskbR]{listPage} function.}

\item{filter}{Filter (one) that should be used in the query. A possible list of pages can be retieved using the \link[pcoskbR]{listFilters} function.}

\item{search_qry}{Searches a keyword in the query.}
}
\value{
A \code{data.frame}. Size and attributes depend on query.
}
\description{
This function scrapes data from PCOSKB. Given a set of filters and corresponding values, it retrieves information about genes, miRNAs, SNPs, diseases, pathways, and ontology associated with PCOS.
}
\examples{
 \code{browse(page = "Genes", filter = "Manually curated", search_qry = "Fem")}
}
\author{
Omar Hassoun
}
