% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneNetworkAnalysis.R
\name{getGeneNetworkAnalysis}
\alias{getGeneNetworkAnalysis}
\title{Gene Network Analysis:  It provides an output of a list of genes with their network properties and tissue specific interactions for each gene.  tab.}
\usage{
getGeneNetworkAnalysis(disease_list, database)
}
\arguments{
\item{disease_list}{character vector containing diseases to be analyzed.}

\item{database}{character vector containing the database from which pathways will be extracted. It can either be "KEGG" or "Reactome".}
}
\value{
a dataframe containing a list of genes with network properties and tissue specific interactions.
}
\description{
Gene Network Analysis:  It provides an output of a list of genes with their network properties and tissue specific interactions for each gene.  tab.
}
