% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennAnalysis.R
\name{getVennAnalysis}
\alias{getVennAnalysis}
\title{Venn Analsis Tool}
\usage{
getVennAnalysis(disease_list, option = "Genes")
}
\arguments{
\item{disease_list}{Diseases for which comorbidity with PCOS will be analyzed. A possible list of diseases can be retrieved using the \code{listDiseases} function.}

\item{option}{It can either be "Genes", "Pathways" or "Ontologies".}
}
\value{
A Venn diagram.
}
\description{
Illustrates shared genes, ontologies or pathways between diseases as 6-way Venn Diagram. It takes a vector of up to 6 diseases and options.
}
\examples{
\code{getVennAnalysis(disease_list = c("Tangier Disease", "Hypercholesterolemia", "Neuropathy" ), option = "Genes")}
}
