% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDiseaseGroup.R
\name{listDiseaseGroup}
\alias{listDiseaseGroup}
\title{List Disease Groups}
\usage{
\code{listDiseaseGroup(dataset = c("Genes", "miRNAs"))}
}
\arguments{
\item{dataset}{Dataset. Either "miRNA" or "Genes".}
}
\value{
A \code{vector} containing disease groups.
}
\description{
Lists disease groups given a dataset.
}
\examples{
\code{listDiseaseGroup(dataset = "Genes")}
}
\author{
Omar Hassoun
}
