% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFilters.R
\name{listFilters}
\alias{listFilters}
\title{lisFilters: Prints the filters for the page inputted.}
\usage{
listFilters(page)
}
\arguments{
\item{page}{It's a string for the page you want to obtain filters to.}
}
\value{
Prints all the filters for the page inputted.
}
\description{
lisFilters: Prints the filters for the page inputted.
}
