% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPages.R
\name{listPages}
\alias{listPages}
\title{listPages}
\usage{
\code{listPages()}
}
\value{
A \code{vector} of pages.
}
\description{
Lists pages available in PCOSKBR2.
}
\examples{
\code{listPages()}
}
\author{
Omar Hassoun
}
