% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneNetworkAnalysis.R
\name{viewInteractions}
\alias{viewInteractions}
\title{View Interactions}
\usage{
\code{viewInteractions(gene, tissue = "all")}
}
\arguments{
\item{gene}{Gene Symbol.}

\item{tissue}{Tissue Type. A possible list of tissues can be vizualised using the \code{listTissues} function.}
}
\value{
Network of of all genes that interact with input gene.
}
\description{
Shows network of interactions of input gene with other genes and vizualise tissue specific interactions.
}
\examples{
\code{viewInteractions(gene = "INS", tissue = "adipose tissue")}
}
\author{
Omar Hassoun
}
